Move server and client code into separate files
